[io] -Fix: do not remove save backup if save has failed
git-svn-id: svn://repo.hu/pcb-rnd/trunk@4336 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/plug_io.c
@@ -803,7 +803,7 @@ int WritePCBFile(const char *Filename, pcb_bool thePcb, const char *fmt)
 
 	fclose(fp);
 	if (fn_tmp != NULL) {
-		if (!conf_core.rc.keep_save_backups)
+		if ((result == 0) && (!conf_core.rc.keep_save_backups))
 			unlink(fn_tmp);
 		free(fn_tmp);
 	}
